Move Results into class
